+++ src/main/java/org/stellar/sdk/responses/sorobanrpc/GetLatestLedgerResponse.java
@@ -1,6 +1,9 @@
 package org.stellar.sdk.responses.sorobanrpc;
 
 import lombok.Value;
+import org.stellar.sdk.Util;
+import org.stellar.sdk.xdr.LedgerCloseMeta;
+import org.stellar.sdk.xdr.LedgerHeader;
 
 /**
  * Response for JSON-RPC method getLatestLedger.
@@ -16,4 +19,30 @@ public class GetLatestLedgerResponse {
   Integer protocolVersion;
 
   Integer sequence;
+
+  Long closeTime;
+
+  /** The field can be parsed as {@link LedgerHeader} object. */
+  String headerXdr;
+
+  /** The field can be parsed as {@link LedgerCloseMeta} object. */
+  String metadataXdr;
+
+  /**
+   * Parses the {@code envelopeXdr} field from a string to an {@link LedgerHeader} object.
+   *
+   * @return the parsed {@link LedgerHeader} object
+   */
+  public LedgerHeader parseHeaderXdr() {
+    return Util.parseXdr(headerXdr, LedgerHeader::fromXdrBase64);
+  }
+
+  /**
+   * Parses the {@code metadataXdr} field from a string to an {@link LedgerCloseMeta} object.
+   *
+   * @return the parsed {@link LedgerCloseMeta} object
+   */
+  public LedgerCloseMeta parseMetadataXdr() {
+    return Util.parseXdr(metadataXdr, LedgerCloseMeta::fromXdrBase64);
+  }
 }